fix: updated image source path in VerificationMessageDisplay.tsx
Fixed the relative path used to display the ai_quiz_logo image in the VerificationMessageDisplay component. This issue was found in the previous code where the image was not displaying due to an incorrect image path. The correct image path has now been updated to "images/ai_quiz_logo.png".