chore: @ts-ignore to bypass TypeScript warning
The option of '@ts-ignore' was introduced in the <Question> component in createQuiz.tsx to bypass an issue with TypeScript warnings. This helps to keep the development process smooth and uninterrupted by any non-critical warning messages. Please note, @ts-ignore is not recommended for use in the production level code and need to be handled appropriately as soon as possible.