feat: added Navbar link to pricing section and hash navigation library
The commit added a smooth hashing navigation link in the navbar to directly navigate to the pricing section of the landing page. This improves the user experience by simplifying the navigation process. Additionally, a new dependency 'react-router-hash-link' was also added for smoothly navigating to any section on the page using hash-link. By providing an ID to the div of the pricing section, the feature takes users directly to the pricing section from the main navigation bar.