feat: added react-confetti library for enhanced visual experience
Added the react-confetti library to the project and used it in the PaymentSuccess component. This library will be used to generate a confetti animation whenever a payment is successfully processed. This will enhance the visual appeal of the application and provide a more positive feedback to the user upon successful transactions.